Include all modules
So far ST has looked at package.json to determine a subset of typings to include. Going forward *all* files will be translated. This solves #463

There are loads of details for computing shorter module names. These will be solved similar to how `exports` was solved in #486, by injecting "proxy modules" into the AST.

Note that some libraries ship typings for multiple hierarchies, typically in folders named `src`, `es`, `es6` and so on. we'll now get all those duplicates as well.
