+++ tests/antd/in/antd/lib/form/Form.d.ts
@@ -1,4 +1,4 @@
-import { FormProps as RcFormProps } from 'rc-field-form/lib/Form';
+import { FormProps as RcFormProps } from 'rc-field-form/es/Form';
 
 export interface FormProps extends Omit<RcFormProps, 'form'> {
     prefixCls?: string;
