+++ tests/commander/check-3/c/commander/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "commander"
-version := "2.15.1-fc4aa5"
+version := "2.15.1-0e0384"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
