+++ tests/firebase-admin/check-3/f/firebase-admin/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "firebase-admin"
-version := "8.2.0-8b474c"
+version := "8.2.0-1cd812"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
