+++ tests/react-integration-test/check-japgolly-3/r/react-dropzone/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "react-dropzone"
-version := "10.1.10-3f6e87"
+version := "10.1.10-76ad43"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
