+++ tests/react-integration-test/check-japgolly-3/r/react-dropzone/src/main/scala/typingsJapgolly/reactDropzone/components/ReactDropzone.scala
@@ -6,10 +6,10 @@ import org.scalajs.dom.HTMLElement
 import typingsJapgolly.StBuildingComponent
 import typingsJapgolly.react.mod.RefAttributes
 import typingsJapgolly.react.mod.global.JSX.Element
-import typingsJapgolly.reactDropzone.mod.DropEvent
-import typingsJapgolly.reactDropzone.mod.DropzoneProps
-import typingsJapgolly.reactDropzone.mod.DropzoneRef
-import typingsJapgolly.reactDropzone.mod.DropzoneState
+import typingsJapgolly.reactDropzone.typingsReactDropzoneMod.DropEvent
+import typingsJapgolly.reactDropzone.typingsReactDropzoneMod.DropzoneProps
+import typingsJapgolly.reactDropzone.typingsReactDropzoneMod.DropzoneRef
+import typingsJapgolly.reactDropzone.typingsReactDropzoneMod.DropzoneState
 import org.scalablytyped.runtime.StObject
 import scala.scalajs.js
 import scala.scalajs.js.annotation.{JSGlobalScope, JSGlobal, JSImport, JSName, JSBracketAccess}
