+++ tests/react-integration-test/check-japgolly-3/r/react-dropzone/src/main/scala/typingsJapgolly/reactDropzone/reactDropzoneStrings.scala
@@ -1,6 +1,6 @@
 package typingsJapgolly.reactDropzone
 
-import typingsJapgolly.reactDropzone.mod.PropTypes
+import typingsJapgolly.reactDropzone.typingsReactDropzoneMod.PropTypes
 import org.scalablytyped.runtime.StObject
 import scala.scalajs.js
 import scala.scalajs.js.annotation.{JSGlobalScope, JSGlobal, JSImport, JSName, JSBracketAccess}
