+++ tests/react-integration-test/check-japgolly-3/s/stardust-ui__react-component-event-listener/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "stardust-ui__react-component-event-listener"
-version := "0.38.0-8cd60e"
+version := "0.38.0-d4b1ad"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
