+++ tests/react-integration-test/check-japgolly-3/s/stardust-ui__react-component-event-listener/src/main/scala/typingsJapgolly/stardustUiReactComponentEventListener/distEsMod.scala
@@ -9,10 +9,9 @@ import org.scalablytyped.runtime.StObject
 import scala.scalajs.js
 import scala.scalajs.js.annotation.{JSGlobalScope, JSGlobal, JSImport, JSName, JSBracketAccess}
 
-/* Inferred short module name */
-object mod {
+object distEsMod {
   
-  @JSImport("@stardust-ui/react-component-event-listener", JSImport.Namespace)
+  @JSImport("@stardust-ui/react-component-event-listener/dist/es", JSImport.Namespace)
   @js.native
   val ^ : js.Any = js.native
   
@@ -20,34 +19,34 @@ object mod {
     
     inline def apply[T /* <: EventTypes */](props: EventListenerOptions[T]): Any = ^.asInstanceOf[js.Dynamic].apply(props.asInstanceOf[js.Any]).asInstanceOf[Any]
     
-    @JSImport("@stardust-ui/react-component-event-listener", "EventListener")
+    @JSImport("@stardust-ui/react-component-event-listener/dist/es", "EventListener")
     @js.native
     val ^ : js.Any = js.native
     
     object defaultProps {
       
-      @JSImport("@stardust-ui/react-component-event-listener", "EventListener.defaultProps")
+      @JSImport("@stardust-ui/react-component-event-listener/dist/es", "EventListener.defaultProps")
       @js.native
       val ^ : js.Any = js.native
       
-      @JSImport("@stardust-ui/react-component-event-listener", "EventListener.defaultProps.capture")
+      @JSImport("@stardust-ui/react-component-event-listener/dist/es", "EventListener.defaultProps.capture")
       @js.native
       def capture: Boolean = js.native
       inline def capture_=(x: Boolean): Unit = ^.asInstanceOf[js.Dynamic].updateDynamic("capture")(x.asInstanceOf[js.Any])
     }
     
-    @JSImport("@stardust-ui/react-component-event-listener", "EventListener.displayName")
+    @JSImport("@stardust-ui/react-component-event-listener/dist/es", "EventListener.displayName")
     @js.native
     def displayName: String = js.native
     inline def displayName_=(x: String): Unit = ^.asInstanceOf[js.Dynamic].updateDynamic("displayName")(x.asInstanceOf[js.Any])
     
-    @JSImport("@stardust-ui/react-component-event-listener", "EventListener.propTypes")
+    @JSImport("@stardust-ui/react-component-event-listener/dist/es", "EventListener.propTypes")
     @js.native
     def propTypes: Capture | Listener = js.native
     inline def propTypes_=(x: Capture | Listener): Unit = ^.asInstanceOf[js.Dynamic].updateDynamic("propTypes")(x.asInstanceOf[js.Any])
   }
   
-  @JSImport("@stardust-ui/react-component-event-listener", "documentRef")
+  @JSImport("@stardust-ui/react-component-event-listener/dist/es", "documentRef")
   @js.native
   val documentRef: RefHandle[
     /* import warning: transforms.QualifyReferences#resolveTypeRef many Couldn't qualify Node */ Any
@@ -55,7 +54,7 @@ object mod {
   
   inline def useEventListener[T /* <: /* import warning: LimitUnionLength.leaveTypeRef Was union type with length 91, starting with typingsJapgolly.stardustUiReactComponentEventListener.stardustUiReactComponentEventListenerStrings.waiting, typingsJapgolly.stardustUiReactComponentEventListener.stardustUiReactComponentEventListenerStrings.error, typingsJapgolly.stardustUiReactComponentEventListener.stardustUiReactComponentEventListenerStrings.abort */ Any */](options: EventListenerOptions[T]): Unit = ^.asInstanceOf[js.Dynamic].applyDynamic("useEventListener")(options.asInstanceOf[js.Any]).asInstanceOf[Unit]
   
-  @JSImport("@stardust-ui/react-component-event-listener", "windowRef")
+  @JSImport("@stardust-ui/react-component-event-listener/dist/es", "windowRef")
   @js.native
   val windowRef: RefHandle[
     /* import warning: transforms.QualifyReferences#resolveTypeRef many Couldn't qualify Node */ Any
