+++ tests/react-integration-test/check-japgolly-3/s/stardust-ui__react-component-ref/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "stardust-ui__react-component-ref"
-version := "0.38.0-af16f8"
+version := "0.38.0-303839"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
