+++ tests/react-integration-test/check-japgolly-3/s/stardust-ui__react-component-ref/src/main/scala/typingsJapgolly/stardustUiReactComponentRef/distEsMod.scala
@@ -1,64 +1,63 @@
-package typingsSlinky.stardustUiReactComponentRef
+package typingsJapgolly.stardustUiReactComponentRef
 
-import slinky.core.ReactComponentClass
-import slinky.core.facade.ReactRef
-import typingsSlinky.stardustUiReactComponentRef.anon.Children
-import typingsSlinky.stardustUiReactComponentRef.anon.InnerRef
-import typingsSlinky.stardustUiReactComponentRef.distEsTypesMod.RefProps
+import japgolly.scalajs.react.facade.React.RefHandle
+import typingsJapgolly.react.mod.FunctionComponent
+import typingsJapgolly.stardustUiReactComponentRef.anon.Children
+import typingsJapgolly.stardustUiReactComponentRef.anon.InnerRef
+import typingsJapgolly.stardustUiReactComponentRef.distEsRefFindNodeMod.default
+import typingsJapgolly.stardustUiReactComponentRef.distEsTypesMod.RefProps
 import org.scalablytyped.runtime.StObject
 import scala.scalajs.js
 import scala.scalajs.js.annotation.{JSGlobalScope, JSGlobal, JSImport, JSName, JSBracketAccess}
 
-/* Inferred short module name */
-object mod {
+object distEsMod {
   
-  @JSImport("@stardust-ui/react-component-ref", JSImport.Namespace)
+  @JSImport("@stardust-ui/react-component-ref/dist/es", JSImport.Namespace)
   @js.native
   val ^ : js.Any = js.native
   
-  @JSImport("@stardust-ui/react-component-ref", "Ref")
+  @JSImport("@stardust-ui/react-component-ref/dist/es", "Ref")
   @js.native
-  val Ref: ReactComponentClass[RefProps] = js.native
+  val Ref: FunctionComponent[RefProps] = js.native
   
-  @JSImport("@stardust-ui/react-component-ref", "RefFindNode")
+  @JSImport("@stardust-ui/react-component-ref/dist/es", "RefFindNode")
   @js.native
-  open class RefFindNode ()
-    extends typingsSlinky.stardustUiReactComponentRef.distEsMod.RefFindNode
+  open class RefFindNode () extends default
   /* static members */
   object RefFindNode {
     
-    @JSImport("@stardust-ui/react-component-ref", "RefFindNode")
+    @JSImport("@stardust-ui/react-component-ref/dist/es", "RefFindNode")
     @js.native
     val ^ : js.Any = js.native
     
-    @JSImport("@stardust-ui/react-component-ref", "RefFindNode.displayName")
+    @JSImport("@stardust-ui/react-component-ref/dist/es", "RefFindNode.displayName")
     @js.native
     def displayName: String = js.native
     inline def displayName_=(x: String): Unit = ^.asInstanceOf[js.Dynamic].updateDynamic("displayName")(x.asInstanceOf[js.Any])
     
-    @JSImport("@stardust-ui/react-component-ref", "RefFindNode.propTypes")
+    @JSImport("@stardust-ui/react-component-ref/dist/es", "RefFindNode.propTypes")
     @js.native
     def propTypes: Children | InnerRef = js.native
     inline def propTypes_=(x: Children | InnerRef): Unit = ^.asInstanceOf[js.Dynamic].updateDynamic("propTypes")(x.asInstanceOf[js.Any])
   }
   
-  @JSImport("@stardust-ui/react-component-ref", "RefForward")
+  @JSImport("@stardust-ui/react-component-ref/dist/es", "RefForward")
   @js.native
   open class RefForward ()
-    extends typingsSlinky.stardustUiReactComponentRef.distEsMod.RefForward
+    extends typingsJapgolly.stardustUiReactComponentRef.distEsRefForwardMod.default
   /* static members */
   object RefForward {
     
-    @JSImport("@stardust-ui/react-component-ref", "RefForward")
+    @JSImport("@stardust-ui/react-component-ref/dist/es", "RefForward")
     @js.native
     val ^ : js.Any = js.native
     
-    @JSImport("@stardust-ui/react-component-ref", "RefForward.displayName")
+    @JSImport("@stardust-ui/react-component-ref/dist/es", "RefForward.displayName")
     @js.native
     def displayName: String = js.native
     inline def displayName_=(x: String): Unit = ^.asInstanceOf[js.Dynamic].updateDynamic("displayName")(x.asInstanceOf[js.Any])
     
-    @JSImport("@stardust-ui/react-component-ref", "RefForward.propTypes")
+    @JSImport("@stardust-ui/react-component-ref/dist/es", "RefForward.propTypes")
     @js.native
     def propTypes: Children | InnerRef = js.native
     inline def propTypes_=(x: Children | InnerRef): Unit = ^.asInstanceOf[js.Dynamic].updateDynamic("propTypes")(x.asInstanceOf[js.Any])
@@ -70,15 +69,15 @@ object mod {
     * @param ref An ref object or function
     * @param node A node that should be passed by ref
     */
-  inline def handleRef[N](ref: typingsSlinky.react.mod.Ref[N], node: N): Unit = (^.asInstanceOf[js.Dynamic].applyDynamic("handleRef")(ref.asInstanceOf[js.Any], node.asInstanceOf[js.Any])).asInstanceOf[Unit]
+  inline def handleRef[N](ref: typingsJapgolly.react.mod.Ref[N], node: N): Unit = (^.asInstanceOf[js.Dynamic].applyDynamic("handleRef")(ref.asInstanceOf[js.Any], node.asInstanceOf[js.Any])).asInstanceOf[Unit]
   
   /** Checks that the passed object is a valid React ref object. */
   inline def isRefObject(ref: Any): /* is react.react.RefObject<any> */ Boolean = ^.asInstanceOf[js.Dynamic].applyDynamic("isRefObject")(ref.asInstanceOf[js.Any]).asInstanceOf[/* is react.react.RefObject<any> */ Boolean]
   
-  @JSImport("@stardust-ui/react-component-ref", "refPropType")
+  @JSImport("@stardust-ui/react-component-ref/dist/es", "refPropType")
   @js.native
   val refPropType: /* import warning: transforms.QualifyReferences#resolveTypeRef many Couldn't qualify PropTypes.Requireable<React.Ref<any>> */ Any = js.native
   
   /** Creates a React ref object from existing DOM node. */
-  inline def toRefObject[T /* <: /* import warning: transforms.QualifyReferences#resolveTypeRef many Couldn't qualify Node */ Any */](node: T): ReactRef[T] = ^.asInstanceOf[js.Dynamic].applyDynamic("toRefObject")(node.asInstanceOf[js.Any]).asInstanceOf[ReactRef[T]]
+  inline def toRefObject[T /* <: /* import warning: transforms.QualifyReferences#resolveTypeRef many Couldn't qualify Node */ Any */](node: T): RefHandle[T] = ^.asInstanceOf[js.Dynamic].applyDynamic("toRefObject")(node.asInstanceOf[js.Any]).asInstanceOf[RefHandle[T]]
 }
