+++ tests/react-integration-test/check-slinky-3/r/react-dropzone/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "react-dropzone"
-version := "10.1.10-438eca"
+version := "10.1.10-606391"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
