+++ tests/react-integration-test/check-slinky-3/r/react-dropzone/src/main/scala/typingsSlinky/reactDropzone/reactDropzoneStrings.scala
@@ -1,6 +1,6 @@
 package typingsSlinky.reactDropzone
 
-import typingsSlinky.reactDropzone.mod.PropTypes
+import typingsSlinky.reactDropzone.typingsReactDropzoneMod.PropTypes
 import org.scalablytyped.runtime.StObject
 import scala.scalajs.js
 import scala.scalajs.js.annotation.{JSGlobalScope, JSGlobal, JSImport, JSName, JSBracketAccess}
