+++ tests/react-integration-test/check-slinky-3/s/stardust-ui__react-component-event-listener/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "stardust-ui__react-component-event-listener"
-version := "0.38.0-f0cd1b"
+version := "0.38.0-1815f0"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
