+++ tests/react-integration-test/check-slinky-3/s/stardust-ui__react-component-ref/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "stardust-ui__react-component-ref"
-version := "0.38.0-615fad"
+version := "0.38.0-927790"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
