+++ tests/vue/check-3/v/vue-scrollto/build.sbt
@@ -1,12 +1,12 @@
 organization := "org.scalablytyped"
 name := "vue-scrollto"
-version := "2.7-d31dd1"
+version := "2.7-e1d6db"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
   "com.olvind" %%% "scalablytyped-runtime" % "2.4.2",
   "org.scalablytyped" %%% "std" % "0.0-unknown-ddd099",
-  "org.scalablytyped" %%% "vue" % "2.5.13-597b3c")
+  "org.scalablytyped" %%% "vue" % "2.5.13-29e5d0")
 publishArtifact in packageDoc := false
 scalacOptions ++= List("-encoding", "utf-8", "-feature", "-language:implicitConversions", "-language:higherKinds", "-language:existentials", "-no-indent", "-source:future")
 licenses += ("MIT", url("http://opensource.org/licenses/MIT"))
