+++ tests/vue/check-3/v/vue/build.sbt
@@ -1,6 +1,6 @@
 organization := "org.scalablytyped"
 name := "vue"
-version := "2.5.13-597b3c"
+version := "2.5.13-29e5d0"
 scalaVersion := "3.1.2"
 enablePlugins(ScalaJSPlugin)
 libraryDependencies ++= Seq(
