+++ ts/src/main/scala/org/scalablytyped/converter/internal/ts/TsTreeScope.scala
@@ -539,16 +539,7 @@ object TsTreeScope {
     val modScope = outsideModule / mod
     addAlternative(mod.name, modScope)
     ret += (mod.name -> modScope)
-    mod.comments.cs.foreach {
-      case Marker.ModuleAliases(aliases) =>
-        aliases.foreach { alias =>
-          ret += ((alias, modScope))
-          addAlternative(alias, modScope)
-        }
-      case _ => ()
-    }
   }
-
 }
 
 /**
